Create dragonfly_dev_commands.py
These are commands for creating dragonfly commands faster. Most of them should be CCR. I don't know where you want to put them. The file will require some tweaking since it was written purely for dragonfly not caster. Sorry I don't know exactly how to tweak it.
A couple of the more sophisticated commands are for making mimic and playback commands with dictation. Thanks to DaneSprite for helping a lot with this.