Bug fix: multi-clipboard wasn't initialized. (#730)
* Bug fix: multi-clipboard wasn't initialized.

Fixed a bug where the multi-clipboard was being erased whenever it was
pasted from before being copied into. This was due to failing to
initialize the multi-clipboard on module load.

* Quick and dirty fix for test failure.

This fixes the issue of test failure caused by the clipboard being read
from disk. All configuration file reading and writing is mocked globally
for all tests. This isn't very elegant, but I'd like my changes to be
minimal until a proper refactor of the utility functions is done.
I would appreciate feedback, especially if I've done something awful.