Pass-through text formatting commands
 When I am writing natural language,
for example in comments and READMEs,
I often wish the dictation would allow Dragon to
choose the capitalization method for me.
Sometimes I want to choose the initial letter, however.

As an example consider the phrase,
"My name is Dusty and I like to dictate"

If I dictate this without a caster command,
it comes up the way I want it to,
but I often use either 'Sing' or 'laws'
because I want to control the capitalization of the 1st letter,
or because I want to insert it with other commands in CCR mode.

When I do that to my example sentence,
it s my name and the 1st person 'I.'
The 3 text formatting commands I've added
operate on only the 1st character, but leave the remaining
words is capitalized by Dragon.