Fix settings window crash on complete command on Windows
The crash happens because logging enabled by default for the RPC
server interferes with the client-server communication when the server
runs on pythonw.

In addition, make sure that methods on GUI widgets are invoked on the
main thread.  When RPC server receives a command that manipulates GUI
widgets, it issues an event which handled on the main thread, and then
widget methods are called by the event handler on the main thread.

Disable default logging in every SimpleXMLRPCServer used in other GUI
elements.