add remap_data to snippet_state
motivation: Snippet is supposed to be able to remap extras the appropriate lambda parameters and those final parameter names along with the extra values are stored in the snippet state under the key "extra_data". However for the control grammar to be able to work, we also need to make sure that it knows the renaming that just happened as it will generate all of the variance using the extras from the rules themselves!