switch to recognition observer class and solve  reregistering problem
up until now, to register the call back the function register_post_recognition_callback was used in a way that allowed reregistering the same callback if the grammar reloads! we now subclass the dragonfly recognition observer and add codein order to allow one instance to be to run the call back!