Add command to open line above
I'm not sure if this is generally useful, so feel free to close.

This command adds a character at the end of the current line
and then insert a new line above.

It is symmetric to the `tell` command, which does the same thing,
except
goes below the current line.

both commands are analogous to the o and O commands in vim.

I chose `hark` as it is kind of the inverse of `tell`,
but dragon often hears it as `heart`, so I used both.