Fixed 'again do' speech recognition history conversion for dragonfly playback #833
The errors occur primarily because the during the conversion to a list of strings for playback it assumes one word per string.
DNS recognition history for built-in commands and proper pronouns can have multiple words/symbols as one string.

The assumption in this new code is that the reported history by the speech engine backend is always correct for the dragonfly playback function.