Allow for TomlConfig without configuration path
In general it was expected that self modifying rules Might need a toml Configuration file in order to save their state. However,for rules like the snippet control rule introduced by #775 is not the case and being forced to Specify a configuration file from which they will write and read Whenever it needs to refresh is simply a burden and a waste of additional time. With this solution Such grammars can use None/"" in order to specify that they do not want a configuration file