- HudPrintMessageHandler sends output to hud
- Redirect dragonfly observers and logging to printer.out
The RPC timeout with the hud can interfere with the dragonfly speech recognition loop. This appears as a stutter recognition latency. Exceptions are tracked so this stutter only happens to end user once. An exception occurs if the hud is not available/python 2/running text engine. If an exception in handle_message, print is managed by SimplePrintMessageHandler